[Phase 1] Admin portal ~ done

+++ controllers/category/index.ts
@@ -2,14 +2,12 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { Prisma } from "@prisma/client";
 
 import prisma from "../../utils/prisma";
-import { generateResponse } from "../../utils";
+import { generateResponse, getUser } from "../../utils";
 import { validateCategory, validateImageUpload } from "../../utils/validation";
-import { upload_on_imagekit } from "../../utils/upload";
+import { delete_image_from_imagekit, upload_on_imagekit } from "../../utils/upload";
 
 const getRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const {
-    parent, child, parentId, childLimit, exclude,
-  } = req.query;
+  const { parent, child, parentId, childLimit, exclude } = req.query;
   const search = req.query.search as string;
 
   const options: Prisma.CategoryFindManyArgs = {};
@@ -17,7 +15,7 @@ const getRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
   if (parent) {
     options.where = {
       parentId: {
-        equals: null,
+        isSet: false,
       },
     };
   }
@@ -52,7 +50,6 @@ const getRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   if (exclude) {
-    console.log({ exclude });
     const ids = typeof exclude === "string" ? [exclude] : exclude;
     options.where = {
       ...options.where,
@@ -100,6 +97,12 @@ const getRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
 };
 
 const postRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const user = await getUser(req);
+
+  if (!user?.isAdmin) {
+    return generateResponse("403", "Unauthorized access.", res, { errorMessage: "You're not authorized.", redirect: true });
+  }
+
   const validationResponse = await validateCategory(req.body);
 
   if (validationResponse) {
@@ -111,6 +114,7 @@ const postRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => 
   if (req.file && validateImageUpload(req.file, res)) {
     const response = await upload_on_imagekit(req.file.buffer, req.file.originalname);
     data.image = response.url;
+    data.categoryImageId = response.fileId;
   }
 
   if (req.body.parent) {
@@ -121,12 +125,28 @@ const postRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => 
     };
   }
 
-  const category = await prisma.category.create({ data });
+  const category = await prisma.category.create({
+    data,
+    include: {
+      parentCategory: {
+        select: {
+          id: true,
+          name: true,
+        },
+      },
+    },
+  });
 
-  return generateResponse("200", "Category saved.", res, { category });
+  return generateResponse("200", "Category saved.", res, { data: category });
 };
 
 const patchRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const user = await getUser(req);
+
+  if (!user?.isAdmin) {
+    return generateResponse("403", "Unauthorized access.", res, { errorMessage: "You're not authorized.", redirect: true });
+  }
+
   const categoryId = req.query?.id as string;
 
   const validateResponse = await validateCategory(req.body);
@@ -141,22 +161,68 @@ const patchRequestHandler = async (req: NextApiRequest, res: NextApiResponse) =>
     throw new Error("Category not found.");
   }
 
-  const data = { name: req.body.name } as Prisma.CategoryUpdateInput;
+  const data = {} as Prisma.CategoryUpdateInput;
 
-  if (req.file && validateImageUpload(req.file, res)) {
-    const response = await upload_on_imagekit(req.file.buffer, req.file.originalname);
-    data.image = response.url;
+  if (req.body.name) {
+    data.name = req.body.name;
   }
 
-  await prisma.category.update({ where: { id: categoryId }, data });
+  if (req.body.parent) {
+    data.parentCategory = {
+      connect: {
+        id: req.body.parent,
+      },
+    };
+  }
 
-  return generateResponse("200", "Category updated.", res);
+  if (req.file && validateImageUpload(req.file, res)) {
+    // → remove old image if exist in records from imagekit server
+    if (category.categoryImageId) {
+      await delete_image_from_imagekit(category.categoryImageId);
+    }
+
+    const response = await upload_on_imagekit(req.file.buffer, req.file.originalname);
+    data.image = response.url;
+    data.categoryImageId = response.fileId;
+  }
+
+  const updatedCategory = await prisma.category.update({
+    where: { id: categoryId },
+    data,
+    include: {
+      parentCategory: {
+        select: {
+          id: true,
+          name: true,
+        },
+      },
+    },
+  });
+
+  return generateResponse("200", "Category updated.", res, { data: updatedCategory });
 };
 
 const deleteRequestHandler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const user = await getUser(req);
+
+  if (!user?.isAdmin) {
+    return generateResponse("403", "Unauthorized access.", res, { errorMessage: "You're not authorized.", redirect: true });
+  }
+
   const categoryId = req.query?.id as string;
 
-  const category = await prisma.category.delete({ where: { id: categoryId } });
+  const category = await prisma.category.findFirst({ where: { id: categoryId } });
+
+  if (!category?.id) {
+    throw new Error("Category not found");
+  }
+
+  // → Delete category image from imagekit server if exist
+  if (category.categoryImageId) {
+    await delete_image_from_imagekit(category.categoryImageId);
+  }
+
+  await prisma.category.delete({ where: { id: categoryId } });
 
   return generateResponse("200", `${category.name} is removed successfylly.`, res);
 };
